Move basp_application_factory to own file
